All the current updates!

+++ src/huc/enum.c
@@ -2,6 +2,7 @@
    All rights reserved.
    See LICENSE for details on use and redistribution. */
 
+#include <stdint.h>
 #include <stdlib.h>
 #include <string.h>
 #include "data.h"
@@ -30,7 +31,7 @@ int define_enum (char *sname, int storage)
 		strcpy(enums[enum_ptr].name, n);
 		/* optional initializer */
 		if (match("=")) {
-			long num;
+			intptr_t num;
 			if (const_expr(&num, ",", "}"))
 				count = num;
 		}
@@ -92,7 +93,7 @@ int find_enum_type (char *name)
 	return (-1);
 }
 
-int find_enum (char *sname, long *val)
+int find_enum (char *sname, intptr_t *val)
 {
 	int i;
